feat: widen block-id storage to sixteen bits
The kernel's ceiling has been sixteen bits for some time; this package
stored block ids one byte per cell and carried its own maximum as a
hardcoded byte-wide literal that did not track the kernel's. Past two
hundred and fifty-six blocks, ids would wrap silently and the mesher
would build wrong geometry for the wrapped ones, with nothing raised.

This is the work a previous change deliberately deferred rather than
overlooked. When the kernel pin moved, its changeset recorded that
re-exporting the kernel's wider ceiling into a byte-wide mask would
corrupt the greedy mesher's packed cell, and kept the old literal as a
stopgap. That reasoning was right, and it is why this change is more
than a type swap.

The greedy mask packs a block id together with ambient-occlusion bits
in one cell and merges while the whole cell repeats. Its shift sat
directly above the old byte-wide id, so widening the storage alone
would have reintroduced exactly the corruption that stopgap avoided,
just at a wider boundary — any id above sixteen thousand three hundred
and eighty-three colliding with the occlusion bits. The shift moves to
sixteen and every buffer backing that mask widens again, because an id
of sixteen bits plus two occlusion bits no longer fits in sixteen.

The packed vertex output carried a second, independent truncation of
the same kind and is widened with it.

The maximum is now re-exported from the kernel rather than restated,
so the two cannot drift apart again — which was the original defect.
Lookup tables keyed by block id were already sized from that constant
and grow on their own; their element types are unchanged. Light grids,
fluid state and rail shapes are deliberately untouched: those really
are one byte per cell and have nothing to do with block ids.

No persisted or wire format exists here — the view is purely in-memory
and the caller copies opaque kernel storage in at the boundary — so
unlike the generator's equivalent change there is no version bump and
no migration to invent.

The existing occlusion regression was pinned at the old byte ceiling
and now uses the real one, which byte storage cannot represent, and a
new test meshes an id that could only ever have been stored as
forty-four and asserts it survives. Two assertions that hardcoded the
old table length were failing against the widened table and are
corrected — they had gone stale rather than staying meaningful.

+++ scripts/bench-meshing.ts
@@ -111,7 +111,7 @@ let sink = 0
 // ---------------------------------------------------------------------------
 
 /**
- * Six linear passes summing a chunk's 65,536 bytes.
+ * Six linear passes summing a chunk's 65,536 cells.
  *
  * The machine-speed reference for the `workloads` ratios. Chosen to share the
  * memory-access character of the six face passes rather than to be a pure ALU
@@ -120,7 +120,7 @@ let sink = 0
  * see the harness header on why workload ratios carry a looser tolerance than
  * guard ratios do.
  */
-const yardstickOver = (blocks: Readonly<Uint8Array>) => (): void => {
+const yardstickOver = (blocks: Readonly<Uint16Array>) => (): void => {
   let total = 0
   for (let pass = 0; pass < FACES.length; pass += 1) {
     for (let index = 0; index < BLOCKS_PER_CHUNK; index += 1) {
@@ -208,7 +208,7 @@ const lookupTableArm = (): void => {
  * the arm below unwraps with the same `AIR` default.
  */
 const getBlockOption = (
-  blocks: Readonly<Uint8Array>,
+  blocks: Readonly<Uint16Array>,
   lx: number,
   y: number,
   lz: number,
@@ -228,7 +228,7 @@ const getBlockOption = (
  * Both walk every cell's neighbour across every face — the same 393,216 reads,
  * including the same 1-cell out-of-bounds shell, that meshing performs.
  */
-const plainWalkArm = (blocks: Readonly<Uint8Array>) => (): void => {
+const plainWalkArm = (blocks: Readonly<Uint16Array>) => (): void => {
   const chunk = { blocks, height: CHUNK_HEIGHT }
   let total = 0
   for (const face of FACES) {
@@ -260,7 +260,7 @@ const plainWalkArm = (blocks: Readonly<Uint8Array>) => (): void => {
  * guard tolerance.
  */
 const frozenGetBlock = (
-  blocks: Readonly<Uint8Array>,
+  blocks: Readonly<Uint16Array>,
   lx: number,
   y: number,
   lz: number,
@@ -272,7 +272,7 @@ const frozenGetBlock = (
   return blocks[y + lz * height + lx * height * CHUNK_SIZE] ?? AIR
 }
 
-const frozenWalkArm = (blocks: Readonly<Uint8Array>) => (): void => {
+const frozenWalkArm = (blocks: Readonly<Uint16Array>) => (): void => {
   let total = 0
   for (const face of FACES) {
     for (let lx = 0; lx < CHUNK_SIZE; lx += 1) {
@@ -286,7 +286,7 @@ const frozenWalkArm = (blocks: Readonly<Uint8Array>) => (): void => {
   sink += total
 }
 
-const optionWalkArm = (blocks: Readonly<Uint8Array>) => (): void => {
+const optionWalkArm = (blocks: Readonly<Uint16Array>) => (): void => {
   let total = 0
   for (const face of FACES) {
     for (let lx = 0; lx < CHUNK_SIZE; lx += 1) {
